Add Flag Mixins
Introduce simple Mixin classes which can be used by resources to
explicitly enable idempotency by simply adding them as the parent
class. Note that these are optional but aim to provide a standardization
of the fields.

Additionally, we've opted to use Mixins rather than Decorators here
because I see decorators as a way to modify behaviour. Corrollary,
Mixins add new behaviour, making them more applicable for what we are
doing.