Create v1 of falcon_idempotency
This commit marks the first "beta" version of falcon_idempotency.
We've introduced two methods of baking in idempotency into your
falcon application.

Simple Solution
---------------
Looks for the existence of `Idempotency-Key` in the request header
which then prompts us to cache the responses for later use.

Comprehensive Solution
----------------------
Provides Mixin classes which can be used as flags to enable the
functionalities for chosen resources.

In addition to these two middlewares, we've introduced a
SentinelException which is used to end the request cycle early.

Squashed Commits
----------------
- Set project in alpha stage (2dbfe64)